4-1 format
